requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
beautifulsoup4==4.7.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
constantly==15.1.0
cryptography==2.6.1
cssselect==1.0.3
Django==2.2
django-crontab==0.7.1
djangorestframework==3.9.3
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
lxml==4.3.3
parsel==1.5.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
python-scrapyd-api==2.1.2
pytz==2019.1
queuelib==1.5.0
requests==2.21.0
Scrapy==1.6.0
scrapyd==1.2.0
service-identity==18.1.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
Twisted==18.9.0
urllib3==1.24.2
w3lib==1.20.0
zope.interface==4.6.0
